refactor: move themes helpers alongside stores
