+++ src/components/molecules/code-block/code-block.astro
@@ -4,7 +4,7 @@ import { Code } from "astro:components";
 import { shikiTheme } from "../../../lib/shiki/theme";
 import { shikiDiffNotation } from "../../../lib/shiki/transformers/diff-notation";
 import { useI18n } from "../../../services/i18n";
-import { SHIKI_SETTING_KEY } from "../../../services/stores/settings";
+import { SHIKI_SETTING_KEY } from "../../../services/stores";
 import { getCSSVars } from "../../../utils/attributes";
 import { InvalidPropsError } from "../../../utils/exceptions";
 import Figure from "../../atoms/figure/figure.astro";
