+++ src/services/stores/index.ts
@@ -1,0 +1,2 @@
+export * from "./stores";
+export * from "./themes";
